Split out contact widget into separate file
